+++ formver.compiler-plugin/core/src/org/jetbrains/kotlin/formver/core/embeddings/callables/NonInlineNamedFunctionSignature.kt
@@ -14,4 +14,3 @@ class NonInlineFunctionSignature(
     override val postconditions: List<ExpEmbedding>,
     override val declarationSource: KtSourceElement?
 ) : CompleteFunctionSignature, NonInlineCallable, NamedFunctionSignature by signature
-
